refactor: reuse logic to fetch history content
